+++ src/kikuchipy/signals/util/_master_pattern.py
@@ -544,7 +544,7 @@ def _project_single_pattern_from_master_pattern(
     # Rotate the detector's view of the crystal
     dc_rotated = rotate_vector(rotation, direction_cosines)
 
-    (nii, nij, niip, nijp, di, dj, dim, djm) = _get_lambert_interpolation_parameters(
+    nii, nij, niip, nijp, di, dj, dim, djm = _get_lambert_interpolation_parameters(
         v=dc_rotated, npx=npx, npy=npy, scale=scale
     )