+++ tests/test_io/test_oxford_h5ebsd.py
@@ -1,4 +1,4 @@
-# Copyright 2019-2024 The kikuchipy developers
+# Copyright 2019-2026 The kikuchipy developers
 #
 # This file is part of kikuchipy.
 #
@@ -16,12 +16,17 @@
 # along with kikuchipy. If not, see <http://www.gnu.org/licenses/>.
 
 import numpy as np
+import pytest
 
 import kikuchipy as kp
+from kikuchipy.io.plugins.oxford_h5ebsd._api import get_binning
 
 
 class TestOxfordH5EBSD:
-    def test_load(
+    @pytest.mark.parametrize(
+        "oxford_h5ebsd_file", ["7.0", "6.0"], indirect=["oxford_h5ebsd_file"]
+    )
+    def test_load_oxford_h5ebsd(
         self, oxford_h5ebsd_file, ni_small_axes_manager, assert_dictionary_func
     ):
         s = kp.load(oxford_h5ebsd_file)
@@ -37,11 +42,23 @@ class TestOxfordH5EBSD:
         # Detector
         det = s.detector
         assert det.pc.shape == (3, 3, 3)
+        assert det.binning == 17.0
         assert np.isclose(det.sample_tilt, 69.9, atol=0.1)
-        assert det.binning == 8
         assert np.isclose(det.tilt, 1.5)
 
     def test_load_unprocessed_patterns(self, oxford_h5ebsd_file):
         s1 = kp.load(oxford_h5ebsd_file)
         s2 = kp.load(oxford_h5ebsd_file, processed=False)
         assert np.allclose(s1.data + 1, s2.data)
+
+    @pytest.mark.parametrize(
+        ["version", "header_group", "binning"],
+        [
+            ("5.0", {"Camera Binning Mode": "8x8 (168x128 px)"}, 8),
+            ("6.0", {"Camera Binning Mode": "Resolution (1244x1024 px)"}, 1),
+            ("6.0", {"Camera Binning Mode": "Speed 3 (156x88 px)"}, 8),
+            ("7.0", {"Camera Mode": "Sensitivity (622x512 px)"}, 2),
+        ],
+    )
+    def test_get_binning(self, version, header_group, binning):
+        assert get_binning(header_group, version) == binning