+++ CHANGELOG.rst
@@ -13,6 +13,23 @@ its best to adhere to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>
 List entries are sorted in descending chronological order. Contributors to each release
 were listed in alphabetical order by first name until version 0.7.0.
 
+0.11.4 (2026-03-04)
+===================
+
+Added
+-----
+- Dependency on the ``packaging`` package.
+  (`#755 <https://github.com/pyxem/kikuchipy/pull/755>`_)
+- Limit version of NumPy < 2.4 to avoid Hough indexing error.
+  (`#755 <https://github.com/pyxem/kikuchipy/pull/755>`_)
+
+Fixed
+-----
+- Reading detector binning factor from Oxford Instruments h5ebsd files (H5OINA) with
+  version v7.0 (AZtec v6.2) should now work.
+  (`#755 <https://github.com/pyxem/kikuchipy/pull/755>`_)
+
+
 0.11.3 (2025-12-17)
 ===================